src/test/java/org/wecancoeit/reviews/ReviewRepositoryTest.java
package org.wecancoeit.reviews;

import org.junit.jupiter.api.Test;

import java.util.Collection;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.assertEquals;

public class ReviewRepositoryTest {

    private ReviewRepository underTest;
    private Review reviewOne = new Review(1L, "Haikyu!!!", "image url", "Anime", "content", "date", "description","citation");
    private Review reviewTwo = new Review(2L, "Demon Slayer", "image url", "Anime", "content", "date", "description","citation");
    private Review reviewThree = new Review(3L, "Seen Deadly Sins", "image url", "Anime", "content", "date", "description","citation");
    private Review reviewFour = new Review(4L, "SailorMoon", "image url", "Anime", "content", "date", "description","citation");

    @Test
    public void shouldFindCReviewOne() {
        underTest = new ReviewRepository(reviewOne);
        Review foundReview = underTest.findOne(1L);
        assertEquals(reviewOne, foundReview);
    }

    @Test
    public void shouldFindReviewOneAndReviewTwp() {
        underTest = new ReviewRepository(reviewOne, reviewTwo);
        Collection<Review> foundReviews = underTest.findAll();
        assertThat(foundReviews).contains(reviewOne, reviewTwo);
    }

}